+++ torchbearer/callbacks/early_stopping.py
@@ -71,6 +71,9 @@ class EarlyStopping(Callback):
 
     def step(self, state):
         current = get_metric('Early Stopping', state, self.monitor)
+        if current is None:
+            return
+
         if self.monitor_op(current - self.min_delta, self.best):
             self.best = current
             self.wait = 0