+++ CHANGELOG.md
@@ -3,6 +3,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+
+## [Unreleased]
+### Added
+### Changed
+### Deprecated
+### Removed
+### Fixed
+- Fixed a bug in access metrics function and callbacks that use it
+- Fixed bug where schedulers were called before optimisers 
+- Fixed a bug where the csv logger closed the file too early
+
 ## [0.5.3] - 2020-01-31
 ### Added
 - Method in bases to access metrics